refactor(control): split reload warm coordination
